requirements.txt
langchain
langchain-google-genai
pypdf
tiktoken
python-docx
python-dotenv
fastapi
uvicorn
